Renamed component create-a-scenario
