+++ src/app/features/selectable-presentation/components/selectable-presentation-form/selectable-presentation-form.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { WalletLayoutComponent } from '@app/core/layout/wallet-layout/wallet-layout.component';
 import { RouterModule } from '@angular/router';
 import { SharedModule } from '@app/shared/shared.module';
-import { CreateAScenarioComponent } from './create-a-scenario.component';
+import { SelectablePresentationFormComponent } from './selectable-presentation-form.component';
 import { HttpClientModule } from '@angular/common/http';
 import { HelperCborSelectableService } from '../../services/helper-cbor-selectable.service';
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -11,8 +11,8 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 
 describe('CBOR CreateAScenarioComponent', () => {
-  let component: CreateAScenarioComponent;
-  let fixture: ComponentFixture<CreateAScenarioComponent>;
+  let component: SelectablePresentationFormComponent;
+  let fixture: ComponentFixture<SelectablePresentationFormComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -26,14 +26,14 @@ describe('CBOR CreateAScenarioComponent', () => {
         SharedModule,
         MatCheckboxModule,
       ],
-      declarations: [ CreateAScenarioComponent ],
+      declarations: [ SelectablePresentationFormComponent ],
       providers: [
         HelperCborSelectableService
       ]
     })
     .compileComponents();
 
-    fixture = TestBed.createComponent(CreateAScenarioComponent);
+    fixture = TestBed.createComponent(SelectablePresentationFormComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
