+++ src/app/features/selectable-presentation/models/Constraint.ts
