+++ src/app/features/selectable-presentation/models/FormSelectableField.ts
@@ -1,4 +1,4 @@
-import { FieldConstraint } from '@features/siop-custom/models/FieldConstraint';
+import { FieldConstraint } from '@features/selectable-presentation/models/FieldConstraint';
 
 export type FormSelectableField = {
   id: number,
