+++ src/app/features/selectable-presentation/models/InputDescriptors.ts
