+++ src/app/features/selectable-presentation/models/Presentation.ts
