+++ src/app/features/selectable-presentation/services/helper-cbor-selectable.service.ts
